reorganize s3 contents retrieval functionalities
